Add Menu (quebra de termos aceitos table)
