+++ front/manager-prov/package.json
@@ -20,6 +20,7 @@
     "@angular/router": "~16.2.0",
     "@po-ui/ng-code-editor": "^16.8.0",
     "@po-ui/ng-components": "^16.8.0",
+    "@po-ui/ng-templates": "^16.8.0",
     "rxjs": "~7.5.5",
     "tslib": "^2.3.0",
     "zone.js": "~0.13.1"
