+++ front/manager-prov/src/app/path/to/home/home.component.html
@@ -1,0 +1,1 @@
+<p>home works!</p>
