+++ front/manager-prov/src/app/path/to/manager/manager.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-manager',
+  templateUrl: './manager.component.html',
+  styleUrls: ['./manager.component.css']
+})
+export class ManagerComponent {
+
+}
